+++ app/inputs/array_input.rb
@@ -1,15 +1,15 @@
 # frozen_string_literal: true
 
-class ArrayInput < SimpleForm::Inputs::StringInput
+class ArrayInput < SimpleForm::Inputs::TextInput
   def input(wrapper_options)
     input_html_options[:type] ||= input_type
     existing_value = Array(object.public_send(attribute_name)).map.with_index do |array_el, index|
-      @builder.text_field(nil, input_html_options.merge(value: array_el,
+      @builder.text_area(nil, input_html_options.merge(value: array_el,
                                                         name: "#{object_name}[#{attribute_name}][]",
                                                         id: "#{object_name}_#{attribute_name}_#{index}"))
     end
     number = Array(object.public_send(attribute_name)).length
-    existing_value.push @builder.text_field(nil,
+    existing_value.push @builder.text_area(nil,
                                              input_html_options.merge(value: nil,
                                                                       name: "#{object_name}[#{attribute_name}][]",
                                                                       id: "#{object_name}_#{attribute_name}_#{number}"))
@@ -17,6 +17,6 @@ class ArrayInput < SimpleForm::Inputs::StringInput
   end
 
   def input_type
-    :email
+    :text
   end
 end