+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_07_07_142820) do
+ActiveRecord::Schema.define(version: 2020_07_09_100253) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -391,6 +391,16 @@ ActiveRecord::Schema.define(version: 2020_07_07_142820) do
     t.string "payment_model_url"
     t.string "pricing_url"
     t.string "security_contact_email", default: "", null: false
+    t.string "resource_geographic_locations", default: [], array: true
+    t.string "certifications", default: [], array: true
+    t.string "standards", default: [], array: true
+    t.string "open_source_technologies", default: [], array: true
+    t.text "changelog", default: [], array: true
+    t.string "grant_project_names", default: [], array: true
+    t.string "multimedia", default: [], array: true
+    t.string "privacy_policy_url"
+    t.string "use_cases_url", default: [], array: true
+    t.datetime "last_update"
     t.index ["name"], name: "index_services_on_name"
     t.index ["provider_id"], name: "index_services_on_provider_id"
     t.index ["resource_organisation_id"], name: "index_services_on_resource_organisation_id"