+++ spec/factories/access_mode.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :access_mode do
+    sequence(:name) { |n| "access mode #{n}" }
+    sequence(:eid) { |n| "access_mode-#{n}" }
+    sequence(:description) { |n| "Description #{n}" }
+    sequence(:extras) { |n| {} }
+  end
+end